services/campaign-director/requirements.txt
a2a-sdk[http-server]>=0.3.25
httpx>=0.27.0
pydantic>=2.0.0
uvicorn>=0.27.0
langgraph>=0.2.0
langchain>=0.2.0
mlflow[genai]>=3